# services/calculator/openapi.yaml
openapi: 3.0.3
info:
  title: Calculator API
  description: A simple API for basic arithmetic operations
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /add:
    post:
      summary: Add two numbers
      description: Returns the sum of two numbers (a + b)
      operationId: addNumbers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRequest'
            example:
              a: 10
              b: 5
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
              example:
                result: 15
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /multiply:
    post:
      summary: Multiply two numbers
      description: Returns the product of two numbers (a × b)
      operationId: multiplyNumbers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRequest'
            example:
              a: 10
              b: 5
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
              example:
                result: 50
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /subtract:
    post:
      summary: Subtract two numbers
      description: Returns the difference of two numbers (a - b)
      operationId: subtractNumbers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationRequest'
            example:
              a: 10
              b: 5
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
              example:
                result: 5
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check
      description: Returns OK if the service is healthy
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
                example: OK

components:
  schemas:
    OperationRequest:
      type: object
      required:
        - a
        - b
      properties:
        a:
          type: number
          format: double
          description: First operand
        b:
          type: number
          format: double
          description: Second operand

    OperationResponse:
      type: object
      properties:
        result:
          type: number
          format: double
          description: Result of the operation

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message